# swagger.yaml
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Work Manager
  description: >-
    Customizable work item manager that leverages Camunda workflows (BPMN and DMN), Form.io and customizable data types
    to model online transaction processing.
servers:
  - url: /api/v1
paths:
...
  /enumerations/{enumerationId}:
    get:
      summary: Get a enumeration from a given id.
      operationId: getEnumerationsById
      tags: [ 'enums' ]
      parameters:
        - in: path
          name: enumerationId
          description: The enumeration id to search on.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The requested enumerations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnumerationResponseModel'
components:
  schemas:
...
    EnumerationResponseModel:
      type: array
      items:
        oneOf:
          - $ref: '#/components/schemas/EnumsModel'
          - $ref: '#/components/schemas/RankedEnumsModel'
    EnumsModel:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
    RankedEnumsModel:
      allOf:
        - $ref: '#/components/schemas/EnumsModel'
        - type: object
          properties:
            rank:
              type: integer
              description: 'Ranking of the enum'
...